update font and refactor language modal
