+++ src/App/App.scss
@@ -491,7 +491,7 @@ button.App-connect-wallet:hover {
 }
 
 .Page-description {
-  color: rgba(255, 255, 255, 0.7);
+  color: #b7b7bd;
   line-height: 2.15rem;
   font-size: 1.5rem;
 }
