+++ src/components/Glp/GlpSwap.css
@@ -150,7 +150,7 @@
   margin: 0 auto;
   justify-content: center;
   min-height: 4.8rem;
-  font-size: 1.5rem;
+  font-size: 1.4rem;
 }
 
 .token-table {
