+++ src/components/NetworkDropdown/NetworkDropdown.tsx
@@ -9,17 +9,17 @@ import language24Icon from "img/ic_language24.svg";
 import settingsIcon from "img/ic_settings_16.svg";
 import arbitrumIcon from "img/ic_arbitrum_24.svg";
 import avaxIcon from "img/ic_avalanche_24.svg";
-import checkedIcon from "img/ic_checked.svg";
 import { importImage } from "lib/legacy";
-import { defaultLocale, dynamicActivate, isTestLanguage, locales } from "lib/i18n";
+import { defaultLocale } from "lib/i18n";
 import { LANGUAGE_LOCALSTORAGE_KEY } from "config/localStorage";
+import LanguageModalContent from "./LanguageModalContent";
 
-const LANGUAGE_MODAL_KEY = "LANGUAGE";
-const NETWORK_MODAL_KEY = "NETWORK";
+const LANGUAGE_MODAL_KEY: string = "LANGUAGE";
+const NETWORK_MODAL_KEY: string = "NETWORK";
 
 export default function NetworkDropdown(props) {
   const currentLanguage = useRef(localStorage.getItem(LANGUAGE_LOCALSTORAGE_KEY) || defaultLocale);
-  const [activeModal, setActiveModal] = useState(null);
+  const [activeModal, setActiveModal] = useState<string | null>(null);
 
   function getModalContent(modalName) {
     switch (modalName) {
@@ -178,35 +178,6 @@ function NetworkMenuItems({ networkOptions, selectorLabel, onNetworkSelect }) {
   });
 }
 
-function LanguageModalContent({ currentLanguage }) {
-  return Object.keys(locales).map((item) => {
-    const image = !isTestLanguage(item) && importImage(`flag_${item}.svg`);
-    return (
-      <div
-        key={item}
-        className={cx("network-dropdown-menu-item  menu-item language-modal-item", {
-          active: currentLanguage.current === item,
-        })}
-        onClick={() => {
-          if (!isTestLanguage(item)) {
-            localStorage.setItem(LANGUAGE_LOCALSTORAGE_KEY, item);
-          }
-          dynamicActivate(item);
-        }}
-      >
-        <div className="menu-item-group">
-          <div className="menu-item-icon">
-            {isTestLanguage(item) ? "🫐" : <img className="network-dropdown-icon" src={image} alt={locales[item]} />}
-          </div>
-          <span className="network-dropdown-item-label menu-item-label">{locales[item]}</span>
-        </div>
-        <div className="network-dropdown-menu-item-img">
-          {currentLanguage.current === item && <img src={checkedIcon} alt={locales[item]} />}
-        </div>
-      </div>
-    );
-  });
-}
 function NetworkModalContent({ networkOptions, onNetworkSelect, selectorLabel, setActiveModal, openSettings }) {
   async function handleNetworkSelect(option) {
     await onNetworkSelect(option);
