+++ src/components/Tab/Tab.css
@@ -1,7 +1,7 @@
 .Tab.block {
   display: grid;
   grid-auto-flow: column;
-  font-size: 1.5rem;
+  font-size: 1.4rem;
   border-radius: 3px;
   overflow: hidden;
   color: rgba(255, 255, 255, 0.7);
@@ -20,7 +20,6 @@
   padding-left: 1.5rem;
   padding-right: 1.5rem;
   cursor: pointer;
-  /* font-size: 1.5rem; */
 }
 
 .Tab.block .Tab-option:hover {
@@ -43,7 +42,7 @@
   cursor: pointer;
   display: inline-block;
   margin-right: 1.5rem;
-  font-size: 1.4rem;
+  font-size: 1.5rem;
 }
 
 .Tab.inline .Tab-option:hover {
