+++ src/pages/Exchange/Exchange.css
@@ -258,7 +258,7 @@
 }
 
 .ExchangeChart-main-price {
-  font-size: 1.5rem;
+  font-size: 1.7rem;
 }
 
 .ExchangeChart-title {
@@ -502,6 +502,7 @@ table.Exchange-list tr:hover {
 
 table.Exchange-list tr.Exchange-list-header {
   background: none;
+  font-size: 1.6rem;
 }
 
 table.Exchange-list tr.Exchange-list-header th {
